Support the GL_COLOR_MATERIAL option
This option makes it so that the material's color is not taken from the
value given to glMaterial(), but from the vertex data. As a matter of
fact, opengx was doing this by default, but that's not correct, because
according to the documentation this behaviour should be disabled by
default.

This was discovered running tutorial 11 from
https://www.swiftless.com/opengltuts.html, because our square was
painted red when in the desktop OpenGL implementation it is white.

Other changes have been made in this commit to bring the lighting
implementation as close as possible to the specifications: the global
ambient light is now correctly included in the computation (before this
commit, it was completely excluded in some code paths, and applied twice
in others).