Add support for emission lighting for materials
Since the on the first TEV stage the parameters A, B and C are unused,
we can use A for the emission lighting: we store the emission color in a
TEV registry, then tell GX_SetTevColorIn() to read the input from that
registry.

This was tested using the tutorial 12 from swiftless.com.