Lighting: remove pre-multiplication with material colors
All these multiplications are already done by the TEV in
setup_render_scene(). The only exception to this is specular color, but
multiplying it with the light color is wrong anyway (we produce a weaker
light) and, most importantly, this light is already combined with the
material's ambient color by GX (see the previous commit's explanation).