Rename exported symbols with _ogx prefix
Exporing symbols with the _ogx_ prefix greatly reduces the risk of
naming collisions with the application's code.

Besides the renaming, move the texture_list variable as a member of the
main structure, to remove one exported symbols.
Use a macro to avoid renaming all occurrences (though at some point
we might want to do that).