Remove undesired external symbols
Completely remove the DirectDrawSurface export methods, opengx is not an
imaging library and does not need these methods itself.

Mark a bunch of internally used functions as static.

Rename the only needed external function to have the `_ogx_` prefix, in
order to avoid collisions with the application's code.

Afther these changes, the symbols exported by libopengx.a (except all
those starting with `gl*`) are the following:

- ogx_initialize (which is meant to be public)
- _ogx_log_level (internal variable)
- _ogx_state (internal variable)
- _ogx_convert_rgb_image_to_DXT1 (internal function)