Don't pass vertex color to GX if not explicitly provided
It is possible to build a display list consisting only of a sequence of
glBegin() / glVertex3f()+ / glEnd() with no color data. If we store the
color information taken from the active color, then during playback the
list would use that color, and not the color which is active at the time
the list is played back.
Therefore, we add a flag to remember whether glColor() was called
between glBegin() and glEnd() and, if it wasn't, then we don't pass the
color data to the GX API.

We could do the same for texture and normal coordinates, but that would
be just an optimization, since there is no way to provide a whole list
of texture or normal coordinates at list playback time. Still, something
to consider.