Add generic converters for most texture formats
Add a Texel base class for the Texel writer classes, so that they can be
used in non-template code. Add a method to set the pixel color from a
GXColor.

Add a DataReaderBase class defining an interface for pixel readers
(using GXColor as the result), and a couple of generic subclasses (one
of which needs to be a template, in order to read pixel components as
bytes, words, longs and floats without excessive pain) to read texture
input data pixel by pixel.

Finally, we combine these two sides inside a generic texture conversion
function, which reads the client pixel data and stores it into a GX
texture.

This is a minimal implementation: we will need to support specifying the
pitch and the pixel alignment. We also do not support loading textures
whose components are encoded with one of the signed types GL_BYTE,
GL_SHORT and GL_INTEGER; these are easy to add, but since they would be
used in the templated reader, and given that they are fairly uncommon,
we leave them out until someone requests them, in order to save on code
size.