Support GL_POLYGON
This can be trivially supported as a triangle fan, since polygons are
guaranteed to be convex and the vertices are specified in the same
order.
The difference will only be noticeable if the rendering mode is set to
GL_LINE, in which case we we'll be producing more lines than needed: the
real GL_POLYGON would draw just the perimeter, while we'll draw some
internal lines too.