projection: adjust matrix using near and far planes
The previous formula only worked if the near plane is close to 1.0, and
it would clip out most of the scene if the near value was less than 0.5
or bigger than 2.0. While the exact reason is still unknown to me, it's
clear that it has to do with the fact that the projection matrix has to
follow a specific layout, and multiplying it for a scale & translate
matrix causes the 3rd element of the bottom row to change from the
expected value of -1.0.

So, let's take another approach: we now build the matrix by reusing the
GL matrix and applying a couple of changes to it:
- transposing it (since it does not mach GX's layout)
- recomputing the elements for the far and near planes according to GX's
  formulas

In practice, this means that we need to extract the near and far values
from the GL matrix, after which we compute the coefficients according to
the formulas used in glFrustum() and glOrtho().