utils: add a function to setup 2D pixel-perfect drawing
When implementing the GL_SELECT render mode and the stencil buffer (and
possibly other operations) we will need to copy the current EFB contents
somewhere, do some operations on the EFB, and finally copy the saved
contents back. We will need all these operations to perfectly operate on
the EFB pixels, and therefore a 2D projection matrix will come handy.

For the time being, we can use this function in our glClear()
implementation.