Implement GL_SELECT render mode
The selection render mode is implemented as follows: when entering the
GL_SELECT mode, we:

- Save the current Z-buffer into a texture
- Set the Z-compare function to GX_ALWAYS
- Disable color and alpha updates

During the selection render mode operation, we also disable glClear and
the swap buffers operation.

When leaving the GL_SELECT mode, we:

- Restore the previously saved Z-buffer
- Re-enable color and alpha updates

These changes have been tested with the gl-117 game, which uses
selection mode in its menu screen.