texture: support generation of texture coordinates
For the time being support is limited to the GL_EYE_LINEAR and
GL_OBJECT_LINEAR modes, and to S and T coordinates. Also, due to
limitation of the GX API, the same generation mode is applied to both
coordinates (that is, the actual setting on the T coordinate is
ignored, and the one from S is used on T too).