texture: don't require a call to glBindTexture()
Surprising as it may be, callint glBindTexture() is not necessary,
because the standard specifies that by default the texture 0 should be
bound.
Up until now, all the programs we have tested with opengx were calling
glBindTexture() both when building the texture and when drawing a frame:
this guaranteed that GX_LoadTexObj() was always called before rendering,
so the texture was available to the TEV. However, generally we cannot
count on this execution pattern, so if we want to make sure that the
texture is ready, we need to load it when preparing the TEV stage.