efb: enable color updates when restoring EFB
We should not rely on the fact that the _ogx_efb_restore_texobj()
function will only e called when color updates are enabled -- or, at the
very least, if we accept this requirement we should document it. It's
more robust if we always enable color updates within the function and
set the dirty bit, as we do for all other GX settings that we modify.