lists: pick current color and normal, if not given
This fixes wrong shading in the dinoshade sample[1]: the program draws
a dinosaur using call lists in which the normal coordinates are not
given; instead, normals are set using glNormal() at the time when the
list is executed, which is something we did not support before, since
all our vertex data was encoded in the GX call lists.

Now, if the color or normal is not given, encode them as indexed data,
which will point to the current color and normal at the time when the
list is executed.

[1]: https://www.opengl.org/archives/resources/code/samples/glut_examples/examples/dinoshade.c