pixels: Reuse the Texel object throughout the conversion
Do not instantiate the Texture object for a single pixel, but just once
at the beginning of the texture creation. This does not save us any
memory allocations (as a matter of fact, the Texel object is allocated
on the stack at no cost), but allows us to implement some optimizations
in the conversion process.