arrays: allow using positions and normals as texture coordinate inputs
The GX_VA_TEX* formats only allow up to two components in the texture
input coordinates (s and t), but OpenGL applications can provide up to
four, along with a texture matrix to transform them.
While we are not able to support this general case (not easily, at
least!), we can handle the case where the array with the texture input
coordinates is the same passed for the positional or normal attributes,
because in that case we can skip re-uploading these coordinates at all
and can setup the TEV to read texture coordinates from GX_TG_POS or
GX_TG_NRM and transform them with the given matrix.

This works well with Neverball, which passes the same array of positions
as texture coordinates when it needs to apply the ball shadow to the
object lying under the ball.