texture: primitive implementation of GL_SPHERE_MAP
This implementation comes with a big simplification: while the OpenGL
specifications say that the texture coordinates should be computed as a
reflection from the point of view of the camera, we do it as if the
camera rays were all parallel to each other (that is, as if the camera
was located infinitely far away). This approximation is acceptable for
objects like spheres or with other curved surfaces, it works pretty
badly with flat surfaces where all the normals have the same value,
because in that case the same texel from the reflection texture gets
applied over to the whole object.